+++ package.json
@@ -12,16 +12,16 @@
         "precommit": "lint-staged"
     },
     "dependencies": {
-        "@material-ui/core": "^4.12.3",
-        "@types/history": "^4.7.3",
-        "i18next": "^20.3.5",
+        "@material-ui/core": "4.12.3",
+        "@types/history": "4.7.3",
+        "i18next": "20.3.5",
         "i18next-http-backend": "1.3.0",
         "prettier-eslint": "13.0.0",
         "prettier-eslint-cli": "5.0.1",
         "react": "17.0.1",
         "react-i18next": "11.11.4",
         "react-native": "0.64.2",
-        "react-native-i18n": "^2.0.15",
+        "react-native-i18n": "2.0.15",
         "redux": "4.1.1",
         "redux-thunk": "2.3.0"
     },
@@ -31,7 +31,7 @@
         "@react-native-community/eslint-config": "2.0.0",
         "@types/jest": "26.0.23",
         "@types/react-native": "0.64.5",
-        "@types/react-native-i18n": "^2.0.0",
+        "@types/react-native-i18n": "2.0.0",
         "@types/react-test-renderer": "16.9.2",
         "@typescript-eslint/eslint-plugin": "4.29.0",
         "babel-eslint": "10.1.0",
@@ -54,7 +54,7 @@
         "metro-react-native-babel-preset": "0.64.0",
         "prettier": "1.19.1",
         "react-test-renderer": "17.0.1",
-        "typescript": "^4.3.5"
+        "typescript": "4.3.5"
     },
     "resolutions": {
         "@types/react": "17"
