+++ src/App.tsx
@@ -2,22 +2,10 @@ import React from 'react';
 import {
   View,
   Text,
-  StyleSheet,
 } from 'react-native';
 import { useTranslation } from 'react-i18next';
 import './utils/i18n/index';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'flex-start',
-    paddingTop: 20,
-  },
-  text: {
-    color: '#808080',
-  },
-});
+import styles from './styles';
 
 const App: React.FunctionComponent = () => {
   const { t } = useTranslation();
