+++ src/styles.ts
@@ -1,0 +1,18 @@
+import {
+  StyleSheet,
+} from 'react-native';
+
+const lightGray = '#808080';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'flex-start',
+    paddingTop: 20,
+  },
+  text: {
+    color: lightGray,
+  },
+});
+export default styles;
